docs: adding .gitignore inside golfsim servlet project
