+++ Software/LMSourceCode/ImageProcessing/golfsim_tomee_webapp/.gitignore
@@ -1,0 +1,4 @@
+
+.vscode
+.idea
+target
