+++ Software/LMSourceCode/ImageProcessing/golfsim_tomee_webapp/src/main/webapp/WEB-INF/web.xml
@@ -4,14 +4,14 @@
          xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee http://xmlns.jcp.org/xml/ns/javaee/web-app_4_0.xsd"
          version="4.0">
 
-  <display-name>hello</display-name>
+  <display-name>PiTrac!</display-name>
   <welcome-file-list>
     <welcome-file>index.html</welcome-file>
   </welcome-file-list>
   <servlet>
     <description></description>
     <servlet-name>MonitorServlet</servlet-name>
-    <servlet-class>com.verdanttechs.jakarta.ee9.MonitorServlet</servlet-class>
+    <servlet-class>com.verdanttechs.jakarta.ee9.servlet.MonitorServlet</servlet-class>
   </servlet>
   <servlet-mapping>
     <servlet-name>MonitorServlet</servlet-name>
